Finish Block 3
Block 3 is finished, both back and frontend.

This commit comes with the complete template, with django snippets to
have as much blocks and questions as the survey has.

It also implements JS to get the sum of the questions in real time.

Soon, this values will be sent to a DB.

v0.4.0