Create DB side B and migrate to PostgresQL
In this commit I created the users, with a way to save the scores with
the survey they did. This comes with the testing necesary to be sure it
all works as it should.
Also, I migrated to PostgresQL as the database system to publish the app
to the web.

v0.4.1