Finish Block 4
This commit finishes block 4. It makes the graph responsive, edits font
sizes, and adds many more styling to the graph and results.
It adds the button to continue to block 5, though it still does not has
a link. After adding this link, it will be ready to ship to production.

v0.6.2